requirements.txt
discord.py==2.3.2
yt-dlp==2023.12.30
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
pydantic==2.5.2
pydantic-settings==2.1.0